fix(rmm): WIP: fix MISRA C 2012 violations
Fixes MISRA C 2012 Rules 2.1, 2.2, 8.3, 8.4,
9.1, 10.3, 10.6, 14.2, 16.5, 21.2, 21.15 and
21.18 violations.

+++ lib/xlat/src/xlat_tables_core.c
@@ -51,7 +51,7 @@ static inline uint64_t *xlat_table_get_empty(struct xlat_ctx *ctx)
 {
 	assert(ctx->tbls->next_table < ctx->tbls->tables_num);
 
-	return &ctx->tbls->tables[XLAT_TABLE_ENTRIES *
+	return &ctx->tbls->tables[(size_t)XLAT_TABLE_ENTRIES *
 					ctx->tbls->next_table++];
 }
 
@@ -242,6 +242,7 @@ static action_t xlat_tables_map_region_action(const struct xlat_mmap_region *mm,
  *
  * NOTE: This function violates misra-c2012-17.2 due to recursivity.
  */
+/* coverity[misra_c_2012_rule_17_2_violation:SUPPRESS] */
 static uintptr_t xlat_tables_map_region(struct xlat_ctx *ctx,
 					struct xlat_mmap_region *mm,
 					uintptr_t table_base_va,
